♻️ Re : 코드 정리
로그인 함수와 로그아웃 함수 분리
